// src/models/response.models.ts
export declare interface ScreenshotResponse {
  type: string;
  text?: string;
  screenshotPath?: string;
  message?: string;
}

export declare interface TranslateResponse {
  type: string;
  screenshotPath?: string;
  message?: string;
}

export declare interface AddTranslationResponse {
  type: string;
  message?: string;
}